Fix declaration of sizes for read and write port

+++ include/arch/x86/io.h
@@ -2,8 +2,8 @@
 #define IO_H
 #include <stdint.h>
 
-void write_port(uint8_t value, uint16_t port);
+void write_port(uint16_t value, uint16_t port);
 
-uint8_t read_port(uint16_t port);
+uint16_t read_port(uint16_t port);
 
 #endif
